requirements.txt
Babel==2.9.1
beautifulsoup4==4.9.3
feedparser==6.0.2
lxml==4.6.3
PyYAML==5.4.1
requests==2.25.1
aiohttp==3.7.4.post0

# dev
flake8
autopep8
coverage

# test
pytest-mock==3.6.1
pytest==6.2.4
Faker==8.8.1
